refactor: improvement coverage collector code
